requirements-txt.txt
google-generativeai>=0.3.0
Pillow>=9.0.0
selenium>=4.1.0
requests>=2.27.1
